Added NUnit tests to project based on Excel Added Excel file with 2-4

+++ Insurance/Customer.cs
@@ -8,9 +8,9 @@ namespace Insurance
 {
     public class Customer
     {
-        public float CalcPremium(int age, string gender)
+        public double CalcPremium(int age, string gender)
         {
-            float premium;
+            double premium;
 
             if (gender == "female")
             {
